Smart Card based e-Purse to Pay change in TSRTC.
Smart Card based payments are gaining popularity in many of the
countries. Smart Card is a convenient way to pay for tickets in RTC.
Debit Cards/Credit Cards require Internet for payment but it will not be
available during travel. To solve the problem of change such as coins in
transactions we are proposing e-Purse to maintain prepaid balance in the
smart card. When the traveller purchases a ticket he can opt for change
payment through e-Purse. The conductor should have hand held smart card
reader/writer to update balance in the   e-Purse of the traveller.